requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.1
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1
selenium==4.15.0
webdriver-manager==4.0.1
feedparser==6.0.10
newspaper3k==0.2.8
tweepy==4.14.0
linkedin-api==2.0.0
sec-api==1.0.32
alpha-vantage==2.3.1
yfinance==0.2.18
openai==0.28.1
anthropic==0.7.7
langchain==0.0.335
chromadb==0.4.15
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
plotly==5.17.0
dash==2.14.1
streamlit==1.27.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.21
alembic==1.12.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.0
asyncio==3.4.3
aiohttp==3.8.6
websockets==11.0.3
schedule==1.2.0
apscheduler==3.10.4
loguru==0.7.2
structlog==23.1.0
prometheus-client==0.17.1
sentry-sdk==1.35.0 